requirements.txt
# FastAPI and web framework
fastapi
uvicorn[standard]

# Environment variables
python-dotenv

# Database
supabase

# Authentication and security
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-auth
google-auth-oauthlib
authlib
itsdangerous

# HTTP client for Supabase
httpx

# Redis for caching
redis

# Pydantic for data validation
pydantic[email]
pydantic-settings

# Testing
pytest
pytest-asyncio

# Image processing (for future use)
Pillow

# Async support
asyncpg

# File handling
aiofiles